requirements.txt
asgiref==3.3.4
celery==5.1.2
Django==3.2.4
djangorestframework==3.12.4
django-celery-results==2.2.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-redis==5.0.0
djoser==2.1.0
inflection==0.5.1
psycopg2-binary==2.8.5
pytz==2021.1
pyyaml==5.4.1
rauth==0.7.3
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
